add validation for user data Relates #61
