+++ src/firebase/firestore/user/EditUserData.js
@@ -1,0 +1,29 @@
+import { db } from '../../config';
+import editUserSchema from '../../../utils/validation';
+
+const EditUserData = async (payload) => {
+  const payloadObj = {
+    id: payload.id,
+    name: payload.name,
+    phoneNumber: payload.phoneNumber,
+    image: payload.image,
+  };
+  try {
+    const { id, name, phoneNumber, image } = await editUserSchema.validate(
+      payloadObj
+    );
+    const response = db.collection('users').doc(id);
+    await response.update({
+      phoneNumber,
+      name,
+      image,
+    });
+    return {
+      data: (await response.get()).data(),
+      msg: 'user updated successfully',
+    };
+  } catch (err) {
+    return err;
+  }
+};
+export default EditUserData;
