--- src/firebase/user/index.js
@@ -1,3 +1,0 @@
-import EditUserData from './EditUserData';
-
-export default EditUserData;
